requirements.txt
python-dotenv
openai
beautifulsoup4
requests
